4connerwithsegment/requirements.txt
requests==2.22.0
Keras==2.3.1
Flask==1.0.2
glob2==0.7
opencv_python==4.1.2.30
imutils==0.5.3
tensorflow==2.0.0
numpy==1.16.3
Pillow==7.0.0
